chore: moved `GradeComponent` schema from `Reporting` to `Grading` context
